+++ scripts/src/form/confirmation/payment.tsx
@@ -3,6 +3,7 @@ import PaymentTable from "./PaymentTable";
 import PaypalClassic from "./PaypalClassic";
 import CCAvenue from "./CCAvenue";
 import ManualApproval from "./ManualApproval";
+import Redirect from "./Redirect";
 import sanitize from "../../sanitize";
 import ExpressionParser from "../../common/ExpressionParser";
 import { IPaymentProps, IPaymentInfo } from "../interfaces";
@@ -11,7 +12,8 @@ let Components = {
   paypal_classic: PaypalClassic,
   ccavenue: CCAvenue,
   manual_approval: ManualApproval,
-  manual_approval_2: ManualApproval
+  manual_approval_2: ManualApproval,
+  redirect: Redirect
 };
 
 class Payment extends React.Component<IPaymentProps, any> {
@@ -23,9 +25,13 @@ class Payment extends React.Component<IPaymentProps, any> {
     };
   }
 
-  getPaymentMethods() {
+  getPaymentMethods(fullyPaid) {
     let paymentMethods = Object.keys(this.props.paymentMethods);
     return paymentMethods.map(paymentMethod => {
+      // If fullyPaid is true, only gets payment methods that apply regardless of whether
+      // one has fully paid (in this case, the "redirect" payment method).
+      if (fullyPaid && paymentMethod !== "redirect") return;
+
       var MyComponent = Components[paymentMethod];
       if (!MyComponent) return;
 
@@ -98,7 +104,7 @@ class Payment extends React.Component<IPaymentProps, any> {
     }
     return (
       <div>
-        <h1 className="text-center my-2">
+        <h1 className="text-center my-2 p-4">
           {this.props.paymentInfo.paymentInfoTableTitle || "Payment"}
         </h1>
         <div>
@@ -161,15 +167,18 @@ class Payment extends React.Component<IPaymentProps, any> {
                   }}
                 />
               )}
-              {this.getPaymentMethods()}
+              {this.getPaymentMethods(true)}
             </div>
           )}
-          {this.props.paymentInfo_owed.total == 0 && (
-            <div className="alert alert-info">
-              We have already received your payment. No additional payment
-              necessary -- you will receive a confirmation email shortly about
-              your update.
-            </div>
+          {this.props.paymentInfo_owed.total === 0 && (
+            <>
+              <div className="alert alert-info">
+                We have already received your payment. No additional payment
+                necessary -- you will receive a confirmation email shortly about
+                your update.
+              </div>
+              {this.getPaymentMethods(false)}
+            </>
           )}
         </div>
       </div>